Move tests into appropriate files
